# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
paths:
...
  /consumptions/{childCode}:
    get:
      tags:
        - Consumptions
      description: Reports all the pending consumptions grouped by year-month
      operationId: getChildConsumptions
      parameters:
        - name: childCode
          description: the child code to list consumptions, if none, all children ar listed
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/YearMonthConsumptionsDTO'
        '404':
          description: Not Found
  /consumptions:
    get:
      tags:
        - Consumptions
      description: Reports all the pending consumptions grouped by year-month
      operationId: getConsumptions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/YearMonthConsumptionsDTO'
    post:
      tags:
        - Consumptions
      description: Sets the consumptions
      operationId: setConsumptions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetYearMonthConsumptionsDTO'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/YearMonthConsumptionsDTO'
...
components:
  schemas:
...
    YearMonthConsumptionsDTO:
      description: Pending consumptions for a year-month
      type: object
      properties:
        yearMonth:
          type: string
        grossAmount:
          type: number
          format: double
        children:
          type: array
          items:
            $ref: '#/components/schemas/ChildConsumtionDTO'
      required:
        - yearMonth
        - grossAmount
        - children
    ChildConsumtionDTO:
      description: Pending consumption of a child for a year-month
      type: object
      properties:
        code:
          type: integer
          format: int32
        shortName:
          type: string
        grossAmount:
          type: number
          format: double
        consumptions:
          type: array
          items:
            $ref: '#/components/schemas/ConsumtionDTO'
      required:
        - code
        - shortName
        - grossAmount
        - consumptions
    ConsumtionDTO:
      description: Consumption item of a child
      properties:
        productId:
          type: string
        units:
          type: number
          format: double
        grossAmount:
          type: number
          format: double
        note:
          type: string
      required:
        - productId
        - units
        - grossAmount
    SetYearMonthConsumptionsDTO:
      description: Pending consumptions for a year-month
      type: object
      properties:
        yearMonth:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/SetChildConsumtionDTO'
      required:
        - yearMonth
        - children
    SetChildConsumtionDTO:
      description: Pending consumption of a child for a year-month
      type: object
      properties:
        code:
          type: integer
          format: int32
        consumptions:
          type: array
          items:
            $ref: '#/components/schemas/SetConsumtionDTO'
      required:
        - code
        - consumptions
    SetConsumtionDTO:
      description: Consumption item of a child
      properties:
        productId:
          type: string
        units:
          type: number
          format: double
        note:
          type: string
      required:
        - productId
        - units